Sun, Oct 1, 2023, 4:27 PM +03:00

+++ population.c
@@ -4,10 +4,10 @@
 
 
 int main(void){
-    int starting_number = get_int("How many lammas at start ");
-    int ending_number = get_int("How many lammas at end ");
+    float starting_number = get_int("How many lammas at start ");
+    float ending_number = get_int("How many lammas at end ");
 
-    int years_to_get_to_goal = ending_number * 12 / (13 * starting_number);
+    float years_to_get_to_goal = ending_number * 12 / (13 * starting_number);
 
 printf(years_to_get_to_goal);
